+++ salt/output/nested.py
@@ -30,7 +30,8 @@ from numbers import Number
 # Import salt libs
 import salt.output
 from salt.ext.six import string_types
-from salt.utils import get_colors, sdecode
+from salt.utils import get_colors
+import salt.utils.locales
 
 
 class NestDisplay(object):
@@ -62,7 +63,7 @@ class NestDisplay(object):
         try:
             return fmt.format(indent, color, prefix, msg, endc, suffix)
         except UnicodeDecodeError:
-            return fmt.format(indent, color, prefix, sdecode(msg), endc, suffix)
+            return fmt.format(indent, color, prefix, salt.utils.locales.sdecode(msg), endc, suffix)
 
     def display(self, ret, indent, prefix, out):
         '''
